// frontend/plugins/welcome/src/api/models/EntDisease.ts
/* tslint:disable */
/* eslint-disable */
/**
 * SUT SA Example API Patient
 * This is a sample server for SUT SE 2563
 *
 * The version of the OpenAPI document: 1.0
 * Contact: support@swagger.io
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
import {
    EntDiseaseEdges,
    EntDiseaseEdgesFromJSON,
    EntDiseaseEdgesFromJSONTyped,
    EntDiseaseEdgesToJSON,
} from './';

/**
 * 
 * @export
 * @interface EntDisease
 */
export interface EntDisease {
    /**
     * Contagion holds the value of the "Contagion" field.
     * @type {string}
     * @memberof EntDisease
     */
    contagion?: string;
    /**
     * DiseaseName holds the value of the "DiseaseName" field.
     * @type {string}
     * @memberof EntDisease
     */
    diseaseName?: string;
    /**
     * Symptom holds the value of the "Symptom" field.
     * @type {string}
     * @memberof EntDisease
     */
    symptom?: string;
    /**
     * 
     * @type {EntDiseaseEdges}
     * @memberof EntDisease
     */
    edges?: EntDiseaseEdges;
    /**
     * ID of the ent.
     * @type {number}
     * @memberof EntDisease
     */
    id?: number;
}

export function EntDiseaseFromJSON(json: any): EntDisease {
    return EntDiseaseFromJSONTyped(json, false);
}

export function EntDiseaseFromJSONTyped(json: any, ignoreDiscriminator: boolean): EntDisease {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'contagion': !exists(json, 'Contagion') ? undefined : json['Contagion'],
        'diseaseName': !exists(json, 'DiseaseName') ? undefined : json['DiseaseName'],
        'symptom': !exists(json, 'Symptom') ? undefined : json['Symptom'],
        'edges': !exists(json, 'edges') ? undefined : EntDiseaseEdgesFromJSON(json['edges']),
        'id': !exists(json, 'id') ? undefined : json['id'],
    };
}

export function EntDiseaseToJSON(value?: EntDisease | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'Contagion': value.contagion,
        'DiseaseName': value.diseaseName,
        'Symptom': value.symptom,
        'edges': EntDiseaseEdgesToJSON(value.edges),
        'id': value.id,
    };
}